requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
albumentations=1.2.1=pypi_0
asttokens=2.0.5=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.0.0=pypi_0
captum=0.5.0=pypi_0
certifi=2022.6.15=py39h06a4308_0
charset-normalizer=2.0.12=pypi_0
cloudpickle=2.0.0=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
datasets=2.3.2=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.5.1=pypi_0
et_xmlfile=1.1.0=py39h06a4308_0
executing=0.8.3=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.7.1=pypi_0
fonttools=4.29.1=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=pypi_0
fsspec=2022.5.0=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
huggingface-hub=0.8.1=pypi_0
idna=3.3=pypi_0
imageio=2.16.0=pypi_0
imbalanced-learn=0.9.0=pypi_0
imblearn=0.0=pypi_0
importlib-metadata=4.11.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipython=8.2.0=pypi_0
jedi=0.18.1=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
llvmlite=0.38.0=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.2=pypi_0
multiprocess=0.70.13=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pypi_0
numba=0.55.1=pypi_0
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyhd3eb1b0_0
opencv-python=4.5.5.64=pypi_0
opencv-python-headless=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.4.1=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
prompt-toolkit=3.0.29=pypi_0
protobuf=3.20.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=8.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydicom=2.2.2=pypi_0
pygments=2.11.2=pypi_0
pyparsing=3.0.7=pypi_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
pywavelets=1.2.0=pypi_0
pyyaml=6.0=pypi_0
qudida=0.0.4=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.6.2=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rsa=4.8=pypi_0
scikit-image=0.19.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
seaborn=0.11.2=pypi_0
setuptools=58.0.4=py39h06a4308_0
sewar=0.4.5=pypi_0
shap=0.40.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
slicer=0.0.7=pypi_0
sqlite=3.37.2=hc218d9a_0
stack-data=0.2.0=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.2.9=pypi_0
timm=0.6.7=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.12.1=pypi_0
torch=1.13.0.dev20220809+cu113=pypi_0
torchaudio=0.13.0.dev20220809+cu113=pypi_0
torchinfo=1.7.0=pypi_0
torchvision=0.14.0.dev20220809+cu113=pypi_0
torchxrayvision=0.0.35=pypi_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pypi_0
transformers=4.20.1=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.8=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=2.1.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xxhash=3.0.0=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
zipp=3.8.0=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
